坐标系修改、配送数据类型修改、店铺详情添加经纬度

+++ app/src/main/java/com/baidu/iov/dueros/waimai/net/entity/response/OrderPreviewBean.java
@@ -588,7 +588,7 @@ public class OrderPreviewBean {
 
                 private String name;
                 private String info;
-                private int reduceFree;
+                private double reduceFree;
                 private String icon_url;
 
                 public String getName() {
@@ -607,7 +607,7 @@ public class OrderPreviewBean {
                     this.info = info;
                 }
 
-                public int getReduceFree() {
+                public double getReduceFree() {
                     return reduceFree;
                 }
